refactor(judge): rename driver→agent_driver, split selection.py, slim __init__.py
Three readability passes from the high-level review:

- driver.py → agent_driver.py. The class `AgentDriver` already had the
  explicit name; the file didn't. New contributors looking for "where
  does the judge call Claude" now have a discoverable filename.

- selection.py → episode_discovery.py + selectors.py. The single file held
  two unrelated concerns: filtering the episode pool to judge
  (`discover_episodes`, `select_episodes`, `EpisodeRef`,
  `load_episode_record`) and picking related-trajectory neighbours
  (`Selector` Protocol + three selectors). They share only the EpisodeRef
  dataclass; splitting clarifies which one a reader is looking at.

- __init__.py shrunk 157 → 45 lines via `from <submodule> import *`.
  Each submodule already maintains its own __all__; the package-level
  surface is now the union, with no separate __all__ to drift out of
  sync. Same public name set verified — every prior import path resolves.

Verified: lint clean, 1076 unit tests passing, judge_e2e smoke 12/12
green against live Sonnet.
